+++ pyuploadcare/__init__.py
@@ -6,4 +6,13 @@ from pyuploadcare.resources.file_group import FileGroup  # noqa: F401
 from pyuploadcare.resources.file_list import FileList  # noqa: F401
 from pyuploadcare.resources.group_list import GroupList  # noqa: F401
 from pyuploadcare.api.entities import Webhook, ProjectInfo  # noqa: F401
+from pyuploadcare.api.search_entities import (  # noqa: F401
+    DatetimeRange,
+    FileSearchRequest,
+    SearchExact,
+    SearchPhrase,
+    SearchSort,
+    SizeRange,
+    TagsFilter,
+)
 from pyuploadcare.client import Uploadcare  # noqa: F401